Update unit test.
